+++ common/datavalues/src/types/mod.rs
@@ -15,11 +15,15 @@
 mod data_df_type;
 mod data_type;
 mod data_type_coercion;
+mod date_converter;
+mod deserializations;
 mod physical_data_type;
 mod serializations;
 
 pub use data_df_type::*;
 pub use data_type::*;
 pub use data_type_coercion::*;
+pub use date_converter::*;
+pub use deserializations::*;
 pub use physical_data_type::*;
 pub use serializations::*;
