+++ common/datavalues/src/types/serializations/boolean.rs
@@ -14,53 +14,8 @@
 
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_io::prelude::*;
 
 use crate::prelude::*;
-
-pub struct BooleanDeserializer {
-    pub builder: BooleanArrayBuilder,
-}
-
-impl TypeDeserializer for BooleanDeserializer {
-    fn de(&mut self, reader: &mut &[u8]) -> Result<()> {
-        let value: bool = reader.read_scalar()?;
-        self.builder.append_value(value);
-        Ok(())
-    }
-
-    fn de_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
-        for row in 0..rows {
-            let mut reader = &reader[step * row..];
-            let value: bool = reader.read_scalar()?;
-            self.builder.append_value(value);
-        }
-
-        Ok(())
-    }
-
-    fn de_text(&mut self, reader: &[u8]) -> Result<()> {
-        let v = if reader.eq_ignore_ascii_case(b"false") {
-            Some(false)
-        } else if reader.eq_ignore_ascii_case(b"true") {
-            Some(true)
-        } else if reader.eq_ignore_ascii_case(b"null") {
-            None
-        } else {
-            return Err(ErrorCode::BadBytes("Incorrect boolean value"));
-        };
-        self.builder.append_option(v);
-        Ok(())
-    }
-
-    fn de_null(&mut self) {
-        self.builder.append_null()
-    }
-
-    fn finish_to_series(&mut self) -> Series {
-        self.builder.finish().into_series()
-    }
-}
 
 pub struct BooleanSerializer {}
