+++ common/datavalues/src/types/serializations/date_time.rs
@@ -17,76 +17,9 @@ use std::ops::AddAssign;
 
 use chrono::Duration;
 use chrono::NaiveDateTime;
-use chrono::TimeZone;
-use chrono_tz::Tz;
 use common_exception::*;
-use common_io::prelude::*;
-use lexical_core::FromLexical;
-use num::cast::AsPrimitive;
 
 use crate::prelude::*;
-
-pub struct DateTimeDeserializer<T: DFPrimitiveType> {
-    pub builder: PrimitiveArrayBuilder<T>,
-    pub tz: Tz,
-}
-
-impl<T> TypeDeserializer for DateTimeDeserializer<T>
-where
-    i64: AsPrimitive<T>,
-    T: DFPrimitiveType,
-    T: Unmarshal<T> + StatBuffer + FromLexical,
-    DFPrimitiveArray<T>: IntoSeries,
-{
-    fn de(&mut self, reader: &mut &[u8]) -> Result<()> {
-        let value: T = reader.read_scalar()?;
-        self.builder.append_value(value);
-        Ok(())
-    }
-
-    fn de_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
-        for row in 0..rows {
-            let mut reader = &reader[step * row..];
-            let value: T = reader.read_scalar()?;
-            self.builder.append_value(value);
-        }
-        Ok(())
-    }
-
-    fn de_text(&mut self, reader: &[u8]) -> Result<()> {
-        if reader.eq_ignore_ascii_case(b"null") {
-            self.builder.append_null();
-            return Ok(());
-        }
-
-        match lexical_core::parse::<T>(reader) {
-            Ok(v) => {
-                self.builder.append_value(v);
-                Ok(())
-            }
-            Err(_) => {
-                let v = std::str::from_utf8(reader)
-                    .map_err_to_code(ErrorCode::BadBytes, || "Cannot convert value to utf8")?;
-                let res = self
-                    .tz
-                    .datetime_from_str(v, "%Y-%m-%d %H:%M:%S%.f")
-                    .map_err_to_code(ErrorCode::BadBytes, || {
-                        "Cannot parse value to DateTime type"
-                    })?;
-                self.builder.append_value(res.timestamp().as_());
-                Ok(())
-            }
-        }
-    }
-
-    fn de_null(&mut self) {
-        self.builder.append_null()
-    }
-
-    fn finish_to_series(&mut self) -> Series {
-        self.builder.finish().into_series()
-    }
-}
 
 pub struct DateTimeSerializer<T: DFPrimitiveType> {
     t: PhantomData<T>,
