+++ common/datavalues/src/types/serializations/mod.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use chrono_tz::Tz;
-use common_exception::ErrorCode;
 use common_exception::Result;
 
 use crate::prelude::*;
@@ -34,61 +32,12 @@ pub use number::*;
 pub use r#struct::*;
 pub use string::*;
 
-pub trait TypeDeserializer: Send + Sync {
-    fn de(&mut self, reader: &mut &[u8]) -> Result<()>;
-    fn de_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()>;
-    /// If error occurrs, append a null by default
-    fn de_text(&mut self, reader: &[u8]) -> Result<()>;
-    fn de_null(&mut self);
-    fn finish_to_series(&mut self) -> Series;
-}
-
 pub trait TypeSerializer: Send + Sync {
     fn serialize_value(&self, value: &DataValue) -> Result<String>;
     fn serialize_column(&self, column: &DataColumn) -> Result<Vec<String>>;
 }
 
 impl DataType {
-    pub fn create_deserializer(&self, capacity: usize) -> Result<Box<dyn TypeDeserializer>> {
-        let data_type = self.clone();
-
-        with_match_primitive_type!(data_type, |$T| {
-                Ok(Box::new(NumberDeserializer::<$T> {
-                    builder: PrimitiveArrayBuilder::<$T>::with_capacity( capacity ),
-                }))
-            },
-
-            {match data_type {
-                DataType::Boolean => Ok(Box::new(BooleanDeserializer {
-                    builder: BooleanArrayBuilder::with_capacity(capacity),
-                })),
-                DataType::Date16 => Ok(Box::new(DateDeserializer::<u16> {
-                    builder: PrimitiveArrayBuilder::<u16>::with_capacity(capacity),
-                })),
-                DataType::Date32 => Ok(Box::new(DateDeserializer::<i32> {
-                    builder: PrimitiveArrayBuilder::<i32>::with_capacity(capacity),
-                })),
-                DataType::DateTime32(tz) => {
-                    let tz = tz.unwrap_or_else(|| "UTC".to_string());
-                    Ok(Box::new(DateTimeDeserializer::<u32> {
-                        builder: PrimitiveArrayBuilder::<u32>::with_capacity(capacity),
-                        tz: tz.parse::<Tz>().unwrap(),
-                    }))
-                }
-                DataType::String => Ok(Box::new(StringDeserializer {
-                    builder: StringArrayBuilder::with_capacity(capacity),
-                })),
-                DataType::Interval(_) => Ok(Box::new(DateDeserializer::<i64> {
-                    builder: PrimitiveArrayBuilder::<i64>::with_capacity(capacity),
-                })),
-                other => Err(ErrorCode::BadDataValueType(format!(
-                    "create_deserializer does not support type '{:?}'",
-                    other
-                ))),
-            }
-        })
-    }
-
     pub fn create_serializer(&self) -> Box<dyn TypeSerializer> {
         match self {
             DataType::Null => Box::new(NullSerializer {}),
